requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
Deprecated==1.2.13
idna==3.2
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
